% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CUSUM_process_Tnk.R
\name{CUSUM_process_Tnk}
\alias{CUSUM_process_Tnk}
\title{CUSUM process for a given time series data}
\usage{
CUSUM_process_Tnk(xt, n, k)
}
\arguments{
\item{xt}{A time sequence.}

\item{n}{Number of observations.}

\item{k}{End time.}
}
\value{
A normalized CUSUM value.
}
\description{
Calculate the cumulative sum (CUSUM) process for a given time sequence \code{xt}.
}
