% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sequence.R
\name{generate_sequence}
\alias{generate_sequence}
\title{Generate a sequence of G values}
\usage{
generate_sequence(phi, lambda, eta)
}
\arguments{
\item{phi}{A numeric value representing the correlation parameter between two contious observations.}

\item{lambda}{A numeric value representing the place where chang point happens.}

\item{eta}{A numeric value representing the constants when generating 'yt' based on 'xt'.}
}
\value{
A vector of G values.
}
\description{
Generate a sequence of G values based on the specified parameters.
}
\examples{
generate_sequence(phi = 0.5, lambda = 0.2, eta = 1)

}
