% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gn_statistics.R
\name{Gn_statistics}
\alias{Gn_statistics}
\title{Calculate the maximum G value for a given time sequence}
\usage{
Gn_statistics(xt, n)
}
\arguments{
\item{xt}{A time series sequence.}

\item{n}{Number of observations.}
}
\value{
The maximum G value.
}
\description{
Calculate the maximum G value for a given time sequence \code{xt} using the specified number of observations \code{n} based on CUSUM process and normalized process.
}
