% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization_process_Vnk.R
\name{Normalization_process_Vnk}
\alias{Normalization_process_Vnk}
\title{Normalized process V with a given time series sequence}
\usage{
Normalization_process_Vnk(xt, n, k)
}
\arguments{
\item{xt}{A time series sequence.}

\item{n}{Number of observations.}

\item{k}{Number of time points in the normalization process.}
}
\value{
The normalized variance.
}
\description{
Normalized process V with a given time series sequence \code{xt} using the specified parameters.
}
