% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SN.R
\name{SN}
\alias{SN}
\title{SN based test}
\usage{
SN(xt, t1, t2)
}
\arguments{
\item{xt}{A time series sequence.}

\item{t1}{Start time point (inclusive).}

\item{t2}{End time point (inclusive).}
}
\value{
The sum of elements in the specified range.
}
\description{
Calculate the sum of elements in the vector \code{xt} from time point \code{t1} to \code{t2}.
}
